chat_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
sqlalchemy>=2.0.25
pymongo>=4.6.0
mysql-connector-python>=8.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.6.0
python-dotenv>=1.0.0
